Extract by-divisor scanner into shared tester
